sorted out doc tests though they require restrucuring
