+++ src/handlers/post.rs
@@ -4,7 +4,7 @@ use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
 
 pub fn handle<'a>(key: KeyEvent, app: &mut App, theme: &'a Theme) {
     match key.code {
-        KeyCode::Esc => app.should_quit = true,
+        KeyCode::Esc => app.stop(),
 
         KeyCode::Tab => {
             app.switch_to_test();
@@ -14,7 +14,7 @@ pub fn handle<'a>(key: KeyEvent, app: &mut App, theme: &'a Theme) {
         KeyCode::Char(c) => {
             if let KeyModifiers::CONTROL = key.modifiers {
                 if c == 'c' {
-                    app.should_quit = true;
+                    app.stop();
                 }
             }
         }
