+++ src/handlers/settings.rs
@@ -8,7 +8,7 @@ pub fn handle<'a>(
     theme: &'a Theme,
 ) {
     match key.code {
-        KeyCode::Esc => app.should_quit = true,
+        KeyCode::Esc => app.stop(),
 
         KeyCode::Tab => {
             app.switch_to_test();
@@ -18,7 +18,7 @@ pub fn handle<'a>(
         KeyCode::Char(c) => {
             if let KeyModifiers::CONTROL = key.modifiers {
                 if c == 'c' {
-                    app.should_quit = true;
+                    app.stop();
                     return;
                 }
             }
@@ -28,7 +28,7 @@ pub fn handle<'a>(
                 'j' => {}
                 'k' => {}
                 'l' => {}
-                'q' => app.should_quit = true,
+                'q' => app.stop(),
                 _ => {}
             }
         }
